Consistenly use uint8_t for monochrome flag
This is a follow-up suggested by Wan-Teh while reviewing a previous
commit, changing to use uint8_t consistently in several structure
for monochrome flag.

+++ aom_scale/yv12config.h
@@ -102,7 +102,7 @@ typedef struct yv12_buffer_config {
   aom_color_primaries_t color_primaries;
   aom_transfer_characteristics_t transfer_characteristics;
   aom_matrix_coefficients_t matrix_coefficients;
-  int monochrome;
+  uint8_t monochrome;
   aom_chroma_sample_position_t chroma_sample_position;
   aom_color_range_t color_range;
   int render_width;
